Struct amethyst_ecs::Planner

pub struct Planner<C> {
    pub systems: Vec<SystemInfo<C>>,
    // some fields omitted
}

System execution planner. Allows running systems via closures, distributes the load in parallel using a thread pool.

Fields
systems: Vec<SystemInfo<C>>

Permanent systems in the planner.

Methods

impl<C> Planner<C> where C: 'static

fn new(world: World, num_threads: usize) -> Planner<C>

Creates a new planner, given the world and the thread count.

fn add_system<S>(&mut self, sys: S, name: &str, priority: i32) where S: 'static + System<C>

Add a system to the dispatched list.

fn run_custom<F>(&mut self, functor: F) where F: 'static + Send + FnOnce(RunArg) -> ()

Runs a custom system.

fn mut_world(&mut self) -> &mut World

Waits for all currently executing systems to finish, and then returns the mutable borrow of the world, allowing to create entities instantly.

fn wait(&mut self)

Waits for all currently executing systems to finish, and then merges all queued changes.

impl<C> Planner<C> where C: Send + 'static + Clone

fn dispatch(&mut self, context: C)

Dispatch all systems according to their associated priorities.

impl<C> Planner<C> where C: 'static

fn run0w1r<R0, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&R0) -> (), R0: Component

impl<C> Planner<C> where C: 'static

fn run0w2r<R0, R1, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&R0, &R1) -> (), R0: Component, R1: Component

impl<C> Planner<C> where C: 'static

fn run0w3r<R0, R1, R2, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&R0, &R1, &R2) -> (), R0: Component, R1: Component, R2: Component

impl<C> Planner<C> where C: 'static

fn run0w4r<R0, R1, R2, R3, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&R0, &R1, &R2, &R3) -> (), R0: Component, R1: Component, R2: Component, R3: Component

impl<C> Planner<C> where C: 'static

fn run1w0r<W0, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0) -> (), W0: Component

impl<C> Planner<C> where C: 'static

fn run1w1r<W0, R0, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &R0) -> (), R0: Component, W0: Component

impl<C> Planner<C> where C: 'static

fn run1w2r<W0, R0, R1, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &R0, &R1) -> (), R0: Component, R1: Component, W0: Component

impl<C> Planner<C> where C: 'static

fn run1w3r<W0, R0, R1, R2, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &R0, &R1, &R2) -> (), R0: Component, R1: Component, R2: Component, W0: Component

impl<C> Planner<C> where C: 'static

fn run1w4r<W0, R0, R1, R2, R3, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &R0, &R1, &R2, &R3) -> (), R0: Component, R1: Component, R2: Component, R3: Component, W0: Component

impl<C> Planner<C> where C: 'static

fn run1w5r<W0, R0, R1, R2, R3, R4, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &R0, &R1, &R2, &R3, &R4) -> (), R0: Component, R1: Component, R2: Component, R3: Component, R4: Component, W0: Component

impl<C> Planner<C> where C: 'static

fn run1w6r<W0, R0, R1, R2, R3, R4, R5, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &R0, &R1, &R2, &R3, &R4, &R5) -> (), R0: Component, R1: Component, R2: Component, R3: Component, R4: Component, R5: Component, W0: Component

impl<C> Planner<C> where C: 'static

fn run1w7r<W0, R0, R1, R2, R3, R5, R6, R7, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &R0, &R1, &R2, &R3, &R5, &R6, &R7) -> (), R0: Component, R1: Component, R2: Component, R3: Component, R5: Component, R6: Component, R7: Component, W0: Component

impl<C> Planner<C> where C: 'static

fn run2w0r<W0, W1, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &mut W1) -> (), W0: Component, W1: Component

impl<C> Planner<C> where C: 'static

fn run2w1r<W0, W1, R0, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &mut W1, &R0) -> (), R0: Component, W0: Component, W1: Component

impl<C> Planner<C> where C: 'static

fn run2w2r<W0, W1, R0, R1, F>(&mut self, functor: F) where F: Send + 'static + FnMut(&mut W0, &mut W1, &R0, &R1) -> (), R0: Component, R1: Component, W0: Component, W1: Component