Struct amethyst_ecs::Storage

pub struct Storage<T, A, D> {
    // some fields omitted
}

A wrapper around the masked storage and the generations vector. Can be used for safe lookup of components, insertions and removes. This is what World::read/write locks for the user.

Methods

impl<T, A, D> Storage<T, A, D>

fn new(alloc: A, data: D) -> Storage<T, A, D>

Create a new Storage

impl<T, A, D> Storage<T, A, D> where A: Deref<Target=Allocator>, D: Deref<Target=MaskedStorage<T>>, T: Component

fn get(&self, e: Entity) -> Option<&T>

Tries to read the data associated with an Entity.

impl<T, A, D> Storage<T, A, D> where A: Deref<Target=Allocator>, D: DerefMut<Target=MaskedStorage<T>>, T: Component

fn get_mut(&mut self, e: Entity) -> Option<&mut T>

Tries to mutate the data associated with an Entity.

fn insert(&mut self, e: Entity, v: T) -> InsertResult<T>

Inserts new data for a given Entity. Returns the result of the operation as a InsertResult<T>

fn remove(&mut self, e: Entity) -> Option<T>

Removes the data associated with an Entity.

fn clear(&mut self)

Clears the contents of the storage.