Struct amethyst_ecs::VecStorage

pub struct VecStorage<T>(_);

Vec-based storage, stores the generations of the data in order to match with given entities. Supposed to have maximum performance for the components mostly present in entities.