Crate amethyst_ecs

Structs

SimBuilder
Consuming builder for easily constructing a new simulations.

Simulation

World
A collection of entities and their respective components.

Traits

Processor
The trait implemented by all processors.

Type Definitions

Entity
An unsigned 64-bit handle to an entity.

ProcessorResult
The error type reported by SimBuilder if they fail to initialize. TODO: original note specified it was en error type reported by a processor, although, as seen below, Processor doesn't have any function to return an error, thus, only SimBuilder can return Result as of now.